+++ src/data/store/pool/seq/impls/mod.rs
@@ -1,0 +1,4 @@
+// devela/src/data/store/pool/seq/impls/mod.rs
+
+mod array; // __pool_seq_impl_array!
+// mod vec; // __pool_seq_impl_vec!